+++ src/include/errorMeasurements.hpp
@@ -7,7 +7,7 @@ class ErrorMeasurementMethod {
 public:
     static double computeVariance(const std::vector<std::vector<std::vector<int>>>& block);
     static double computeMAD(const std::vector<std::vector<std::vector<int>>>& block);
-    static int computeMaxPixelDiff(const std::vector<std::vector<std::vector<int>>>& block);
+    static double computeMaxPixelDiff(const std::vector<std::vector<std::vector<int>>>& block);
     static double computeEntropy(const std::vector<std::vector<std::vector<int>>>& block);
 };
